Add a partially functional GUI
The GUI class can be run directly to show a mockup of the game. The
player's grid (left) shows the different colors. The opponents's grid is
clickable, a dialog will popup showing the coordinates of the clicked
cell.

Currently no way to connect the GUI to the rest of the program, will be
added later.