Add Ship Placement Screen
When starting the game, the player must now place their ships on the
grid. The ships can be rotated. Already placed ships can't be removed.
When all ships are placed, a button to advance to the game proper is
enabled.

The code in GUI.run() looks like ass, but it works. I'll try to clean it
up and add some documentation later.